requirements.txt
asgiref==3.8.1 ; python_version >= "3.12" and python_version < "4.0"
django-environ==0.11.2 ; python_version >= "3.12" and python_version < "4"
django==5.0.4 ; python_version >= "3.12" and python_version < "4.0"
djangorestframework==3.15.1 ; python_version >= "3.12" and python_version < "4.0"
numpy==1.26.4 ; python_version >= "3.12" and python_version < "4.0"
opencv-python==4.9.0.80 ; python_version >= "3.12" and python_version < "4.0"
pillow==10.3.0 ; python_version >= "3.12" and python_version < "4.0"
psycopg2==2.9.9 ; python_version >= "3.12" and python_version < "4.0"
sqlparse==0.4.4 ; python_version >= "3.12" and python_version < "4.0"
tzdata==2024.1 ; python_version >= "3.12" and python_version < "4.0" and sys_platform == "win32"
